Introduce a small abstraction to verify a session

+++ backend/src/api/deleteUser.ts
@@ -1,17 +1,19 @@
 import { Static } from 'runtypes';
 
 import { DeleteUserRequest, DeleteUserResponse } from '../shared/api/schema';
-import { getPool } from '../storage/storage';
-import {
-  sessionLifespanSinceActiveMs,
-  sessionLifespanSinceCreationMs,
-} from '../constants/constants';
+import { getPool, sessionIdToUserId } from '../storage/storage';
 
 export default async function deleteUser(
   payload: Static<typeof DeleteUserRequest>['payload'],
 ): Promise<Static<typeof DeleteUserResponse>['payload']> {
   // Get the database connection pool.
   const pool = await getPool();
+
+  // Fetch the user ID.
+  const userId = await sessionIdToUserId(payload.sessionId);
+  if (userId === null) {
+    return { type: 'NotLoggedIn' };
+  }
 
   // Fetch a client from the pool.
   const client = await pool.connect();
@@ -24,43 +26,12 @@ export default async function deleteUser(
     // being created both happen or neither happens at all.
     await client.query('BEGIN');
 
-    // Fetch the session.
-    const sessions = (
-      await client.query<{
-        createdAt: Date;
-        refreshedAt: Date;
-        userId: string;
-      }>(
-        'SELECT created_at AS "createdAt", refreshed_at AS "refreshedAt", user_id AS "userId" ' +
-          'FROM session ' +
-          'WHERE id = $1 ' +
-          'LIMIT 1',
-        [payload.sessionId],
-      )
-    ).rows;
-    if (sessions.length === 0) {
-      await client.query('ROLLBACK');
-      return { type: 'NotLoggedIn' };
-    }
-    const session = sessions[0];
-
-    // Make sure the session hasn't expired.
-    if (
-      session.createdAt.valueOf() + sessionLifespanSinceCreationMs <=
-        Date.now() ||
-      session.refreshedAt.valueOf() + sessionLifespanSinceActiveMs <= Date.now()
-    ) {
-      return { type: 'NotLoggedIn' };
-    }
-
     // Delete any sessions for the user.
-    await client.query<{}>('DELETE FROM session WHERE user_id = $1', [
-      session.userId,
-    ]);
+    await client.query<{}>('DELETE FROM session WHERE user_id = $1', [userId]);
 
     // Delete any log in invitations for the user.
     await client.query<{}>('DELETE FROM log_in_invitation WHERE user_id = $1', [
-      session.userId,
+      userId,
     ]);
 
     // Fetch the user.
@@ -74,7 +45,7 @@ export default async function deleteUser(
           'FROM "user" ' +
           'WHERE id = $1 ' +
           'LIMIT 1',
-        [session.userId],
+        [userId],
       )
     ).rows[0];
 
@@ -82,7 +53,7 @@ export default async function deleteUser(
     await client.query<{}>(
       'INSERT INTO previous_user_email (user_id, email, normalized_email) ' +
         'VALUES ($1, $2, $3)',
-      [session.userId, user.email, user.normalizedEmail],
+      [userId, user.email, user.normalizedEmail],
     );
 
     // Mark the user as deleted.
@@ -90,7 +61,7 @@ export default async function deleteUser(
       'UPDATE "user" ' +
         'SET email = NULL, normalized_email = NULL, deleted = true ' +
         'WHERE id = $1',
-      [session.userId],
+      [userId],
     );
 
     // Commit the transaction.
